domain-specific/shop/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["warren-wupeng <warren.wu@visionwx.com>"]
readme = "README.md"
packages = [{include = "server", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.0"}
stripe = "^10.7.0"
python-dotenv = "^1.0.1"
httpx = {extras = ["socks"], version = "^0.27.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
